Add project and edit user pages updated for design, use of Input component, and added reset button to each
