+++ src/components/ProjectCard.js
@@ -18,7 +18,7 @@ class ProjectCard extends Component {
         return this.props.projects.map((item,i) => {
             return (
             <div className="col-md-3 card"
-                onClick={() => this.props.history.push('/project/' + item._id)}
+                onClick={() => this.props.history.push('/project/view/' + item._id)}
                 key={i}
                 id={item._id}>
                 <div class="card-body">
@@ -35,7 +35,7 @@ class ProjectCard extends Component {
       }
     }
     render() {
-
+        console.log('render project card', this.props);
          return (
             <div className="container">
                 <div className="row justify-content-center">
