+++ src/components/UserEdit.js
@@ -1,13 +1,22 @@
 import React, { Component } from 'react';
 // import '../stylesheets/components/UserEdit.css';
 import '../stylesheets/main.css'; // for dev
-import Button from './Button.js';
+import Button from './Button';
+import Input from './Input';
 
 
-class UserEdit extends Component {
+class UserEditNew extends Component {
     constructor(props) {
         super(props);
-        this.state = this.props.user;
+        if (this.props.user) {
+            this.state = {
+                username: this.props.user.username,
+                displayName: this.props.user.displayName,
+                avatar: this.props.user.avatar,
+                skillset: this.props.user.skillset,
+                email: this.props.user.email
+            }    
+        }
     }
     componentDidMount() {
         // if user is not logged in and therefore user info is null, redirect to home
@@ -23,20 +32,29 @@ class UserEdit extends Component {
         // for now I am forcing update with each change.
         return true;
     }
-    handleChange = (e) => {
+    handleChange = (name, value) => {
         let obj = {};
-        if (e.target.name === 'skillset') {
-            let arr = e.target.value.split(',');
-            obj[e.target.name] = arr;
+        if (name === 'skillset') {
+            let arr = value.split(',');
+            obj[name] = arr;
         } else {
-            obj[e.target.name] = e.target.value;            
+            obj[name] = value;            
         }
         this.setState(obj);
     }
-    handlePost = (e) => {
+    handleReset = () => {
+        this.setState({
+            username: this.props.user.username,
+            displayName: this.props.user.displayName,
+            avatar: this.props.user.avatar,
+            skillset: this.props.user.skillset,
+            email: this.props.user.email
+        });
+    }
+    handleSubmit = (e) => {
         e.preventDefault();
         let user = {};
-        user._id = this.state._id;
+        user._id = this.props.user._id;
         user.username = this.state.username;
         user.displayName = this.state.displayName;
         user.avatar = this.state.avatar;
@@ -53,52 +71,61 @@ class UserEdit extends Component {
             if (typeof this.state.skillset === 'object') {
                 skillset = this.state.skillset.toString();
             }
+            let inputFields = [
+                {
+                  label: 'Username',
+                  name: 'username',
+                  placeholder: '',
+                  value: this.state.username,
+                  required: true
+                },
+                {
+                  label: 'Display Name',
+                  name: 'displayName',
+                  placeholder: '',
+                  value: this.state.displayName,
+                  required: true
+                },
+                {
+                  label: 'Email',
+                  name: 'email',
+                  placeholder: 'Enter your email address',
+                  value: this.state.email,
+                  required: true
+                },
+                {
+                  label: 'Skillset',
+                  name: 'skillset',
+                  value: skillset,
+                  placeholder: 'Separate values by commas'
+                },
+                {
+                  label: 'Avatar URL',
+                  name: 'avatar',
+                  placeholder: '',
+                  value: this.state.avatar
+                }
+              ];
             return (
                 <div className="container">
                     <div className="row">
                         <div className="col">
-                            <div className='user-edit' >
+                            <div className='material-card' >
                                 <h1>Edit User Profile</h1>
-                                <form>
-                                    <table>
-                                        <tr>
-                                            <td className='table-col-1'>User Name: </td>
-                                            <td className='table-col-2'>
-                                                <input type='text' name='username' value={this.state.username} onChange={this.handleChange}/><br/>
-                                            </td>
-                                        </tr>
-                                        <tr>
-                                            <td className='table-col-1'>Display Name: </td>
-                                            <td className='table-col-2'>
-                                                <input type='text' name='displayName' value={this.state.displayName} onChange={this.handleChange}/><br/>
-                                            </td>
-                                        </tr>
-                                        <tr>
-                                            <td className='table-col-1'>Email: </td>
-                                            <td className='table-col-2'>
-                                                <input type='text' name='email' value={this.state.email} onChange={this.handleChange}/><br/>
-                                            </td>
-                                        </tr>
-                                        <tr>
-                                            <td className='table-col-1'>Avatar URL: </td>
-                                            <td className='table-col-2'>
-                                                <input type='text' name='avatar' value={this.state.avatar} onChange={this.handleChange}/><br/>
-                                            </td>
-                                        </tr>
-                                        <tr>
-                                            <td className='table-col-1'>Skillset: </td>
-                                            <td className='table-col-2'>
-                                                <input type='text' name='skillset' value={skillset} onChange={this.handleChange}/><br/>            
-                                            </td>
-                                        </tr>
-                                    </table>
+                                <form onSubmit={this.handleSubmit}>
+                                    <fieldset>
+                                        {inputFields.map(item => {
+                                            return <Input data={item} onChange={this.handleChange} />
+                                        })}
+                                        <div className='d-flex justify-content-around'>
+                                            <input type='submit' className='btn' value='Submit' />
+                                            <input type='reset' className='btn' value='Reset' onClick={this.handleReset}/>
+                                            <Button label='Cancel' redirect={'/user/view/'+this.props.user._id} />
+                                        </div>
+                                    </fieldset>
                                 </form>
-                                <div className="btn-row">
-                                    <button className='btn btn-primary' onClick={this.handlePost}>Submit</button>
-                                    <button className='btn btn-primary' onClick={() => this.props.history.push('/user/'+this.state._id)}>Cancel</button>
-                                </div>
                             </div>
-                            <button className='btn btn-primary' onClick={() => this.props.history.push('/')}>Back to Home</button>
+                            <Button label='Back to main'/>
                         </div>
                     </div>
                 </div>
@@ -108,4 +135,4 @@ class UserEdit extends Component {
     }
 }
 
-export default UserEdit;
+export default UserEditNew;
