+++ src/index.js
@@ -15,7 +15,6 @@ import ProjectCard from './components/ProjectCard.js';
 import ProjectInfo from './components/ProjectInfo.js';
 import UserInfo from './components/UserInfo.js';
 import UserEdit from './components/UserEdit.js';
-import UserEditNew from './components/UserEditNew.js';
 import AddProject from './components/AddProject.js';
 import Footer from './components/Footer.js';
 import Button from './components/Button.js';
@@ -117,22 +116,22 @@ class App extends Component {
           return <ProjectCard {...routeProps} {...this.state}  />
         }
         }/>
-        <Route path="/project/:id" render={(routeProps)=> {
+        <Route path="/project/view/:id" render={(routeProps)=> {
           return <ProjectInfo {...routeProps} {...this.state} />
         }
         }/>
-        <Route path="/user/:id" render={(routeProps)=> {
+        <Route path="/user/view/:id" render={(routeProps)=> {
           return <UserInfo {...routeProps} {...{user: this.state.user}} />
         }
         }/>
-        <Route path="/editUser" render={(routeProps)=> {
-          return <UserEditNew {...routeProps} {...{
-            user: this.state.user,
-            onUserPost: this.postUser
-          }} />
-        }
+        <Route path="/user/edit/" render={(routeProps)=> {
+            return <UserEdit {...routeProps} {...{
+              user: this.state.user,
+              onUserPost: this.postUser
+            }} />
+          }
         }/>
-        <Route path="/addproject" render={(routeProps)=> {
+        <Route path="/project/add/" render={(routeProps)=> {
               return <AddProject
                 {...routeProps}
                 {...{
