+++ src/stylesheets/components/UserEdit.scss
@@ -14,35 +14,3 @@ body {
     text-transform: uppercase;
     margin-top: 30px;
 }
-
-.btn-row {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-button {
-    margin: 20px 20px 0 20px;
-}
-
-table {
-    width: 100%;
-}
-
-tr {
-    height: 50px;
-}
-
-.table-col-1 {
-    width: 20%;
-}
-
-.table-col-2 {
-    width: 100%;
-}
-
-input {
-    width: 100%;
-    border: none;
-    border-bottom: solid 1px #707070;
-}
