+++ src/stylesheets/components/UserInfo.scss
@@ -15,16 +15,6 @@ body {
     margin-top: 30px;
 }
 
-.btn-row {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-
-button {
-    margin: 20px 20px 0 20px;
-}
-
 td {
     vertical-align: top;
 }
