--- src/stylesheets/components/addProject.scss
@@ -1,11 +1,0 @@
-@import "../base/helpers.scss";
-
-.btn-container {
-    display: flex;
-    justify-content: space-around;
-}
-
-input.btn {
-    margin: auto;
-    width: 150px;
-}
