+++ src/stylesheets/main.scss
@@ -5,7 +5,6 @@
 // FONTS AND TYPOGRAPHY
 @import "base/typography.scss";
 // // COMPONENTS STYLESHEETS
-@import "components/addProject.scss";
 @import "components/ProjectCard.scss";
 @import "components/ProjectInfo.scss";
 @import "components/Nav.scss";
